+++ sdk/src/constants.ts
@@ -1,3 +1,4 @@
+import {TIME} from './from_osx/time';
 import {NetworkDeployment} from './internal';
 import {ProposalMetadata, SupportedNetwork, SupportedVersion} from './types';
 import {activeContractsList} from '@aragon/osx-ethers';
@@ -5,44 +6,6 @@ import {activeContractsList as activeContractsListV1_0_0} from '@aragon/osx-ethe
 import {keccak256} from '@ethersproject/keccak256';
 import {Network} from '@ethersproject/networks';
 import {toUtf8Bytes} from '@ethersproject/strings';
-
-export const TIME = {
-  SECOND: 1,
-  get MINUTE() {
-    return 60 * this.SECOND;
-  },
-  get HOUR() {
-    return 60 * this.MINUTE;
-  },
-  get DAY() {
-    return 24 * this.HOUR;
-  },
-  get WEEK() {
-    return 7 * this.DAY;
-  },
-  get YEAR() {
-    return 365 * this.DAY;
-  },
-
-  toMilliseconds(seconds: number) {
-    return seconds * 1000;
-  },
-  get SECOND_MS() {
-    return this.toMilliseconds(this.SECOND);
-  },
-  get HOUR_MS() {
-    return this.toMilliseconds(this.HOUR);
-  },
-  get DAY_MS() {
-    return this.toMilliseconds(this.DAY);
-  },
-  get WEEK_MS() {
-    return this.toMilliseconds(this.WEEK);
-  },
-  get YEAR_MS() {
-    return this.toMilliseconds(this.YEAR);
-  },
-};
 
 /** Timeout that will be applied to operations involving
  * many fetch requests that could take a long time */
