+++ sdk/src/from_osx/bitmap.ts
