+++ sdk/src/from_osx/permission.ts
