+++ sdk/src/from_osx/proposal.ts
